--- src/app/my/layout.tsx
@@ -1,23 +1,0 @@
-import Sidebar from '../_components/Sidebar';
-import type { Metadata } from 'next';
-
-export const metadata: Metadata = {
-  title: 'Connection | 회원페이지',
-  description: 'Connection 회원페이지',
-};
-
-export default async function MyPageLayout({
-  children,
-}: {
-  children: React.ReactNode;
-}) {
-  return (
-    <section className="grid h-max w-full flex-1 grid-cols-[2fr,1fr] justify-between gap-x-4 bg-sub-color1-transparent px-5 pb-20 pt-4 md:px-16 xl:grid-cols-[1fr,2fr,1fr]">
-      <aside className="hidden w-full rounded-lg bg-white shadow-float xl:flex">
-        <Sidebar />
-      </aside>
-
-      {children}
-    </section>
-  );
-}
