--- src/app/my/manage/myclass/page.tsx
@@ -1,28 +1,0 @@
-'use client';
-import { useState } from 'react';
-import ClassDetailView from './_components/ClassDetailView';
-import ClassListView from './_components/ClassListView';
-
-const MyClassPage = () => {
-  const [activeView, setActiveView] = useState('list');
-
-  const handleItemClick = () => {
-    setActiveView('detail');
-  };
-
-  const handleGoBack = () => {
-    setActiveView('list');
-  };
-
-  return (
-    <>
-      {activeView === 'list' ? (
-        <ClassListView onItemClick={handleItemClick} />
-      ) : (
-        <ClassDetailView onGoBack={handleGoBack} />
-      )}
-    </>
-  );
-};
-
-export default MyClassPage;
