+++ src/app/mypage/instructor/income/_components/AccountInfo.tsx
@@ -11,8 +11,32 @@ interface AccountInfoProps {
 }
 const AccountInfo = ({ view, handleApply }: AccountInfoProps) => {
   return view === 'main' ? (
-    <aside className="flex max-h-[16.5rem] flex-col gap-3">
-      <section className="flex h-fit w-full max-w-xs flex-col rounded-lg bg-white shadow-float shadow-float">
+    <aside className="grid w-full grid-cols-1 gap-x-4 gap-y-3 md:grid-cols-2 lg:grid-cols-1">
+      <section className="flex flex-col rounded-lg bg-white px-5 py-4 shadow-float">
+        <ul className="mb-3 flex w-full flex-col gap-1 whitespace-nowrap text-sm text-gray-100">
+          <li className="flex items-center">
+            <p className="w-10 font-semibold">총 수익</p>
+            <span className="ml-[0.69rem] text-lg font-bold text-black">
+              1,230,800원
+            </span>
+          </li>
+          <li className="flex items-center">
+            <p className="w-10 font-semibold">이번 달</p>
+            <span className="ml-2.5 text-lg font-bold text-sub-color1">
+              1,230,800원
+            </span>
+          </li>
+        </ul>
+        <button
+          onClick={handleApply}
+          className="flex h-7 w-full items-center justify-center rounded-md bg-main-color text-sm font-semibold text-white"
+        >
+          <MoneySVG width="18" height="18" fill="white" stroke="white" />
+          정산신청
+        </button>
+      </section>
+
+      <section className="flex h-fit w-full flex-col rounded-lg bg-white shadow-float shadow-float">
         <h2 className="flex h-12 w-full items-center justify-between border-b border-solid border-gray-700 pl-[1.19rem] text-lg font-semibold text-gray-100">
           계좌 정보
           <BigArrowSVG width="34" height="34" className="fill-gray-500" />
@@ -27,33 +51,9 @@ const AccountInfo = ({ view, handleApply }: AccountInfoProps) => {
           </li>
         </ul>
       </section>
-
-      <section className="flex flex-col rounded-lg bg-white px-5 py-4 shadow-float shadow-float">
-        <ul className="mb-3 flex w-full flex-col gap-1 whitespace-nowrap text-sm text-gray-100">
-          <li className="flex items-center">
-            <p className="w-10 font-semibold">총 수익</p>
-            <span className="ml-[0.69rem] text-lg font-bold text-black">
-              1,230,800원
-            </span>
-          </li>
-          <li className="flex items-center ">
-            <p className="w-10 font-semibold">이번 달</p>
-            <span className="ml-[0.69rem] text-lg font-bold text-sub-color1">
-              1,230,800원
-            </span>
-          </li>
-        </ul>
-        <button
-          onClick={handleApply}
-          className="flex h-7 w-full items-center justify-center rounded-md bg-main-color text-sm font-semibold text-white"
-        >
-          <MoneySVG width="18" height="18" fill="white" stroke="white" />
-          정산신청
-        </button>
-      </section>
     </aside>
   ) : (
-    <aside className="flex h-fit w-full max-w-xs flex-col rounded-lg bg-white p-4 shadow-float">
+    <aside className="flex h-fit w-full flex-col rounded-lg bg-white p-4 shadow-float">
       <h2 className="mb-2 text-sm font-semibold text-main-color">
         *정산금 지급 기준
       </h2>
