+++ src/app/mypage/instructor/income/_components/IncomeRange.tsx
@@ -29,7 +29,7 @@ const IncomeRange = () => {
   ) => {
     if (!fromValue && !toValue && range?.from) {
       setFromValue(format(range.from, 'y-MM-dd'));
-      setToValue(format(range.from, 'y-MM-dd')); // 추가된 부분
+      setToValue(format(range.from, 'y-MM-dd'));
       return;
     }
 
@@ -64,9 +64,9 @@ const IncomeRange = () => {
   };
 
   return (
-    <div className="flex gap-[0.88rem]">
+    <div className="flex gap-3.5">
       <div ref={ref}>
-        <div className="relative flex h-7 w-full max-w-[312px] items-center rounded-md border border-solid border-gray-500 pl-[0.69rem] text-base text-gray-100">
+        <div className="relative flex h-7 w-full max-w-[20rem] items-center rounded-md border border-solid border-gray-500 pl-2.5 text-base text-gray-100">
           <DateInput
             placeholder="시작 날짜"
             value={fromValue || ''}
