+++ src/app/mypage/instructor/income/page.tsx
@@ -1,6 +1,5 @@
 'use client';
 import { useState } from 'react';
-import AccountInfo from './_components/AccountInfo';
 import IncomeTable from './_components/IncomeDataViewer';
 import IncomeOverview from './_components/IncomeOverview';
 import Payment from './_components/Payment';
@@ -17,16 +16,14 @@ const IncomePage = () => {
   };
 
   return (
-    <>
-      <section className="w-full max-w-[40rem]">
-        <div className="flex flex-col gap-4">
-          <IncomeOverview view={view} handlePrev={handlePrev} />
-          {view === 'main' ? <IncomeTable /> : <Payment />}
-        </div>
-      </section>
-
-      <AccountInfo view={view} handleApply={handleApply} />
-    </>
+    <div className="flex w-full flex-col gap-4 px-9 xl:px-0">
+      <IncomeOverview
+        view={view}
+        handlePrev={handlePrev}
+        handleApply={handleApply}
+      />
+      {view === 'main' ? <IncomeTable /> : <Payment />}
+    </div>
   );
 };
