+++ src/app/mypage/instructor/manage/myclass/_components/ClassDetailView.tsx
