+++ src/app/mypage/instructor/manage/myclass/_components/detail/ClassTable.tsx
